requirements.txt
flask==3.0.0
werkzeug==3.0.1
SQLAlchemy==2.0.25
python-dotenv==1.0.0
gunicorn==21.2.0
transformers==4.36.2
torch==2.1.2
numpy==1.26.3
python-telegram-bot==20.7
python-telegram-bot[job-queue]
psycopg2-binary==2.9.9
nltk==3.8.1
requests==2.31.0
sentence-transformers==2.2.2
python-dateutil==2.8.2
pytz==2024.1
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6 